# src/main/resources/openapi/users-api.yaml
openapi: 3.0.3
info:
  title: Users API
  description: Contract-first definition of the springCRUD User API.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
tags:
  - name: users
    description: User and address operations
paths:
  /users/create:
    post:
      tags: [users]
      operationId: createUser
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users/create/async:
    post:
      tags: [users]
      operationId: createUserAsync
      summary: Create a user asynchronously via Kafka (JSON)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '202':
          description: Creation request accepted for async processing

  /users/create/async-proto:
    post:
      tags: [users]
      operationId: createUserAsyncProto
      summary: Create a user asynchronously via Kafka (Protobuf)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '202':
          description: Creation request accepted for async processing

  /users/create-idem:
    post:
      tags: [users]
      operationId: createUserIdem
      summary: Create a user idempotently
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: User created (new request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '200':
          description: User already created (cached idempotent response)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  /users/{id}:
    get:
      tags: [users]
      operationId: getUser
      summary: Get a user by id
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    put:
      tags: [users]
      operationId: updateUser
      summary: Update a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '204':
          description: User updated
    delete:
      tags: [users]
      operationId: deleteUser
      summary: Delete a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted

  /users/{id}/address:
    put:
      tags: [users]
      operationId: updateUserAddress
      summary: Set the current address of a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddressRequest'
      responses:
        '204':
          description: Address updated

  /users/{id}/addresses:
    get:
      tags: [users]
      operationId: getUserAddresses
      summary: List all addresses of a user
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: The user's addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AddressResponse'

components:
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    UserRequest:
      type: object
      required: [email, yearsExperience]
      properties:
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
        phoneNumber:
          type: string
          nullable: true
        yearsExperience:
          type: integer
          format: int32
        address:
          $ref: '#/components/schemas/AddressRequest'
    UserResponse:
      type: object
      required: [id, email, yearsExperience, seniority]
      properties:
        id:
          type: integer
          format: int64
        firstName:
          type: string
          nullable: true
        lastName:
          type: string
          nullable: true
        email:
          type: string
        phoneNumber:
          type: string
          nullable: true
        yearsExperience:
          type: integer
          format: int32
        seniority:
          type: string
          enum: [J, M, S]
        address:
          $ref: '#/components/schemas/AddressResponse'
    AddressRequest:
      type: object
      required: [streetName, streetNumber, city, postcode]
      properties:
        streetName:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        postcode:
          type: string
    AddressResponse:
      type: object
      required: [streetName, streetNumber, city, postcode, isCurrent]
      properties:
        streetName:
          type: string
        streetNumber:
          type: string
        city:
          type: string
        postcode:
          type: string
        isCurrent:
          type: boolean